Added Game test fixture, moved new game IPC test from Board to Game
